Move RQC to MDS
