+++ src/tribler-core/tribler_core/components/metadata_store/remote_query_community/settings.py
