+++ src/tribler-core/tribler_core/config/tribler_config_sections.py
@@ -6,8 +6,8 @@ from tribler_core.components.masterkey.settings import BootstrapSettings, DHTSet
     Ipv8Settings, \
     TrustchainSettings, WatchFolderSettings
 from tribler_core.components.libtorrent.settings import DownloadDefaultsSettings, LibtorrentSettings
+from tribler_core.components.metadata_store.remote_query_community.settings import RemoteQueryCommunitySettings
 from tribler_core.components.popularity.settings import PopularityCommunitySettings
-from tribler_core.modules.remote_query_community.settings import RemoteQueryCommunitySettings
 from tribler_core.components.resource_monitor.settings import ResourceMonitorSettings
 from tribler_core.modules.torrent_checker.settings import TorrentCheckerSettings
 from tribler_core.modules.tunnel.community.settings import TunnelCommunitySettings
